Implemented a fixed time step loop

+++ src/HOE/high_level_components/sprite.cpp
@@ -65,7 +65,7 @@ void Sprite::freeResources()
 void Sprite::Draw(SDL_Renderer** renderer)
 {
     HFMath::Vector2 pos = m_Transform.GetGlobalPosition();
-    SDL_Rect renderRect = { pos.GetX(), pos.GetY(), m_Width, m_Height };
+    SDL_Rect renderRect = { (int)pos.GetX(), (int)pos.GetY(), m_Width, m_Height };
     SDL_RenderCopy(*renderer, m_Texture, NULL, &renderRect);
 }
